+++ packages/console/src/monaco/index.ts
@@ -1,3 +1,4 @@
 export { default as MonacoUtils } from './MonacoUtils';
 export { default as MonacoProviders } from './MonacoProviders';
 export { default as MonacoTheme } from './MonacoTheme.module.scss';
+export * from './MonacoThemeProvider';