+++ packages/embed-chart/src/App.tsx
@@ -42,7 +42,7 @@ async function loadFigure(
  * E.g. http://localhost:3000/?name=myFigure will attempt to open a figure `myFigure`
  * If no query param is provided, it will display an error.
  * By default, tries to connect to the server defined in the VITE_CORE_API_URL variable, which is set to http://localhost:10000/jsapi
- * See create-react-app docs for how to update these env vars: https://create-react-app.dev/docs/adding-custom-environment-variables/
+ * See Vite docs for how to update these env vars: https://vitejs.dev/guide/env-and-mode.html
  */
 function App(): JSX.Element {
   const [model, setModel] = useState<ChartModel>();