+++ playwright-ci.config.ts
@@ -11,6 +11,7 @@ const config: PlaywrightTestConfig = {
     timeout: 60 * 1000,
     reuseExistingServer: false,
   },
+  reporter: [['github'], ['html']],
 };
 
 export default config;